use separate helper file instead of exporting functions from CLI binary script
Ensures binary runs with or without passing it to Node. Removes need to check if binary is run directly or imported.
